requirements.txt
haystack-ai
weaviate-haystack
anthropic-haystack
pypdf
sentence-transformers